ref(github): Move issue tool into plugin tools module
Keep GitHub plugin registration in index.js while moving the typed issue creation tool into a dedicated tools.js module, matching the plugin/tool split used by other packages.

Refs GH-660
