+++ packages/junior-github/package.json
@@ -21,6 +21,7 @@
     "index.d.ts",
     "index.js",
     "permissions.js",
+    "tools.js",
     "skills",
     "SETUP.md"
   ],
